Refactored DecayHandler; Made plugin always use non-generation listeners
