+++ src/main/java/com/redchestraven/food/fooddecay/listeners/TradeForFoodListener.java
@@ -1,11 +1,9 @@
 package com.redchestraven.food.fooddecay.listeners;
 
-import com.redchestraven.food.fooddecay.DecayHandler;
+import com.redchestraven.food.fooddecay.handlers.DecayFoodHandler;
 import com.redchestraven.food.fooddecay.FoodDecay;
 import com.redchestraven.food.fooddecay.consts.CustomDataKeys;
-import com.redchestraven.food.fooddecay.consts.EventNames;
 import org.bukkit.Material;
-import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.inventory.ClickType;
@@ -19,23 +17,20 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.logging.Logger;
 
 public final class TradeForFoodListener implements Listener
 {
 	private static final Logger logger = Logger.getLogger("FoodDecay");
 	private static TradeForFoodListener _tradeForFoodListener = null;
-	private static DecayHandler _decayHandler;
-	private static final Map<String, Boolean> _enabledEvents = new HashMap<>();
+	private static DecayFoodHandler _decayFoodHandler;
 	private static final List<InventoryAction> _pickupActions = List.of(InventoryAction.HOTBAR_SWAP, InventoryAction.PICKUP_ALL,
 			InventoryAction.PICKUP_HALF, InventoryAction.SWAP_WITH_CURSOR, InventoryAction.MOVE_TO_OTHER_INVENTORY);
 	private static final CustomDataKeys cdk = new CustomDataKeys();
 
 	private TradeForFoodListener(JavaPlugin plugin)
 	{
-		_decayHandler = DecayHandler.GetInstance(plugin);
-		UpdateConfig(plugin.getConfig());
+		_decayFoodHandler = DecayFoodHandler.GetInstance(plugin);
 	}
 
 	public static TradeForFoodListener GetInstance(JavaPlugin plugin)
@@ -46,20 +41,10 @@ public final class TradeForFoodListener implements Listener
 		return _tradeForFoodListener;
 	}
 
-	public static void UpdateConfig(FileConfiguration config)
-	{
-		//logger.info("Events after reload: " + _enabledEvents.get(EventNames.onTradeForFood);
-
-		_enabledEvents.clear();
-		_enabledEvents.put(EventNames.onTradeForFood, config.getBoolean(EventNames.onTradeForFood));
-
-		//logger.info("Events after reload: " + _enabledEvents.get(EventNames.onTradeForFood);
-	}
-
 	@EventHandler
 	public void OnTradeForFood(InventoryClickEvent ice)
 	{
-		if(FoodDecay._enabled && _enabledEvents.get(EventNames.onTradeForFood))
+		if(FoodDecay._enabled)
 		{
 			Inventory inventory = ice.getClickedInventory();
 
@@ -70,7 +55,7 @@ public final class TradeForFoodListener implements Listener
 			{
 				logger.info("Click used: " + ice.getAction());
 				ItemStack result = new ItemStack(ice.getCurrentItem());
-				_decayHandler.AddDecayTimeIfDecayingFood(result);
+				_decayFoodHandler.AddDecayTimeIfDecayingFood(result);
 
 				//The above already handles if it's a regular click, so the code below is purely for when it's a shift-click
 				if(result.getItemMeta().getPersistentDataContainer().has(cdk.expirationDate, PersistentDataType.STRING))
@@ -224,7 +209,7 @@ public final class TradeForFoodListener implements Listener
 					else
 					{
 						logger.info("Single trade, so regular handling...");
-						_decayHandler.AddDecayTimeIfDecayingFood(ice.getCurrentItem());
+						_decayFoodHandler.AddDecayTimeIfDecayingFood(ice.getCurrentItem());
 					}
 				}
